backend-fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-firestore
firebase-admin